+++ example/seriesOperation.py
@@ -27,7 +27,7 @@ frame = frame_shape(length=12, width = 5, height = 16)
 
 # make the series of frames
 # appliy the operation in order to make alternating internal rail patern
-frame_series = series(shape=frame, size=5, length_offset=3, operation=rail_operation)
+frame_series = series(shape=frame, size=5, length_offset=3, operation=rail_operation, union=False)
 
 #show_object(frame_series)
 cq.exporters.export(frame_series,'stl/seriesOperation.stl')